Refactor: PathResult와 Path 객체 리팩토링
