+++ src/main/java/nextstep/subway/path/domain/PathResult.java
@@ -3,21 +3,19 @@ package nextstep.subway.path.domain;
 import java.util.Collections;
 import java.util.List;
 import nextstep.subway.station.domain.Station;
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 
 public class PathResult {
 
     private List<Station> stations;
     private int distance;
 
-    private PathResult(List<Station> stations, double distance) {
+    private PathResult(List<Station> stations, int distance) {
         this.stations = stations;
-        this.distance = (int) distance;
+        this.distance = distance;
     }
 
-    public static PathResult of(DijkstraShortestPath dijkstraShortestPath, Station start, Station destination) {
-        return new PathResult(dijkstraShortestPath.getPath(start, destination).getVertexList(),
-            dijkstraShortestPath.getPathWeight(start, destination));
+    public static PathResult of(List<Station> stations, double distance) {
+        return new PathResult(stations, (int) distance);
     }
 
     public List<Station> getStations() {
